# backend/api_gateway/openapi-config.yaml
swagger: '2.0'
info: 
  title: StudentShare Backend API
  description: API gateway to reach the various microservices in the cloud in a bundled way
  version: 1.0.0
host: backend.apigateway.studentshare.cloud.goog
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  jwt:
    flow: "implicit"
    authorizationUrl: ""
    type: "oauth2"
    x-google-issuer: "StudentShare"
    x-google-jwks_uri: "https://storage.googleapis.com/studentshare-key-bucket/StudentSharePublicKeyJWK.json"
    x-google-audiences: "https://www.studentshare.de" 
paths:
...
  /files:
    get:
      tags:
        - files
      summary: Home File Screen, shows every File from a specific major and only if the user has the rights for it.
      operationId: file home
      x-google-backend:
        address: https://student-fileshare-api-b6d727ldhq-ey.a.run.app/files
        jwt_audience: https://student-fileshare-api-b6d727ldhq-ey.a.run.app
      security:
        - jwt: []
      parameters:
        - name: userInfo
          in: body
          description: Information of the User to show him the available files
          schema:
            type: object
            required:
              - token
      responses:
        200:
          description: Available Files to download
    options:
      tags:
        - files
      operationId: file cors
      x-google-backend:
        address: https://student-fileshare-api-b6d727ldhq-ey.a.run.app/files
        jwt_audience: https://student-fileshare-api-b6d727ldhq-ey.a.run.app
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
...
  /files/upload:
    post:
      tags:
        - files
      summary: Upload a file from a user
      operationId: file upload
      x-google-backend:
        address: https://student-fileshare-api-b6d727ldhq-ey.a.run.app/upload
        jwt_audience: https://student-fileshare-api-b6d727ldhq-ey.a.run.app
      security:
        - jwt: []
      parameters:
        - name: formdata
          in: body
          description: The file wrapped in a FormData object
          schema:
            type: object
      responses:
        200:
          description: "Get a status message back"
          schema:
            type: object
    options:
      tags:
        - files
      operationId: file upload cors
      x-google-backend:
        address: https://student-fileshare-api-b6d727ldhq-ey.a.run.app/upload
        jwt_audience: https://student-fileshare-api-b6d727ldhq-ey.a.run.app
      responses:
        200:
          description: "Allow"
        401:
          description: "Cors not allowed"
...